refactor: extract project-connection.js from project.js (PR-13)
Move handleConnection and handleDisconnection into a dedicated module.
Connection handler manages initial state sync, session restore, presence
tracking, and WS event wiring. project.js reduced from 1,358 to 1,191
lines.
